+++ crucible-cli/test/Overrides.hs
@@ -1,13 +1,24 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Overrides where
+module Overrides
+  ( setupOverrides
+  ) where
 
+import qualified Control.Lens as Lens
+import Control.Lens hiding ((:>), Empty)
+import Control.Monad.Except (runExceptT)
 import Control.Monad.IO.Class
+import qualified Control.Monad.State.Strict as State
+import System.IO
+
+import qualified Prettyprinter as PP
+import qualified Prettyprinter.Render.Text as PP
 
 import Data.Parameterized.Context hiding (view)
 import qualified Data.Parameterized.Map as MapF
@@ -17,16 +28,23 @@ import What4.Expr.Builder
 import What4.Interface
 import What4.ProgramLoc
 import qualified What4.Protocol.Online as WPO
+import What4.Solver (LogData(..), defaultLogData)
+import What4.Solver.Z3 (z3Adapter)
 
 import Lang.Crucible.Backend
+import qualified Lang.Crucible.Backend as CB
 import qualified Lang.Crucible.Backend.Online as CBO
+import qualified Lang.Crucible.Backend.Prove as CB
 import qualified Lang.Crucible.Concretize as Conc
 import Lang.Crucible.Types
 import Lang.Crucible.FunctionHandle
 import Lang.Crucible.Simulator
+import qualified Lang.Crucible.Simulator as CS
+import qualified Lang.Crucible.Utils.Seconds as Sec
+import qualified Lang.Crucible.Utils.Timeout as CTO
 
 
--- Some overrides for testing purposes.
+-- | Set up all test overrides
 setupOverrides ::
   ( IsSymBackend sym bak
   , sym ~ ExprBuilder scope st fs
@@ -47,7 +65,11 @@ setupOverrides bak ha =
                      <*> pure (UseOverride (mkOverride "nondetBranchesTest" (nondetBranchesTest (Just sym))))
      f4 <- FnBinding <$> mkHandle ha "concBool"
                      <*> pure (UseOverride (mkOverride "concBool" (concBool bak)))
-     return [(f1, InternalPos),(f2,InternalPos),(f3,InternalPos),(f4,InternalPos)]
+     f5 <- FnBinding <$> mkHandle ha "proveObligations"
+                     <*> pure (UseOverride (mkOverride "proveObligations" (proveObligations (Just sym))))
+     f6 <- FnBinding <$> mkHandle ha "crucible-print-assumption-state"
+                     <*> pure (UseOverride (mkOverride "crucible-print-assumption-state" (printAssumptionState (Just sym))))
+     return [(f1, InternalPos),(f2,InternalPos),(f3,InternalPos),(f4,InternalPos),(f5,InternalPos),(f6,InternalPos)]
 
 
 -- Test the @symbolicBranch@ override operation.
@@ -142,3 +164,46 @@ nondetBranchesTest _proxy =
        , (p2, reg @2 <$> getOverrideArgs, Just (OtherPos "second branch"))
        , (fallbackPred, fallback, Just (OtherPos "default branch"))
        ]
+
+-- | Prove all outstanding proof obligations
+proveObligations :: (IsSymInterface sym, sym ~ (ExprBuilder t st fs)) =>
+  proxy sym ->
+  OverrideSim p sym ext r EmptyCtx UnitType (RegValue sym UnitType)
+proveObligations _proxy =
+  ovrWithBackend $ \bak ->
+  do let sym = backendGetSym bak
+     h <- printHandle <$> getContext
+     liftIO $ do
+       hPutStrLn h "Attempting to prove all outstanding obligations!\n"
+
+       let logData = defaultLogData { logCallbackVerbose = \_ -> hPutStrLn h
+                                    , logReason = "assertion proof" }
+       let timeout = CTO.Timeout (Sec.secondsFromInt 5)
+       let prover = CB.offlineProver timeout sym logData z3Adapter
+       let strat = CB.ProofStrategy prover CB.keepGoing
+       let ppResult o =
+             \case
+               CB.Proved {}  -> unlines ["Proof Succeeded!", show $ ppSimError (proofGoal o ^. labeledPredMsg)]
+               CB.Disproved {} -> unlines ["Proof failed!", show $ ppSimError (proofGoal o ^. labeledPredMsg)]
+               CB.Unknown {} -> unlines ["Proof inconclusive!", show $ ppSimError (proofGoal o ^. labeledPredMsg)]
+       let printer = CB.ProofConsumer $ \o r -> hPutStrLn h (ppResult o r)
+       runExceptT (CB.proveCurrentObligations bak strat printer) >>=
+         \case
+           Left CTO.TimedOut -> hPutStrLn h "Proof timed out!"
+           Right () -> pure ()
+
+       clearProofObligations bak
+
+-- | Print the current assumption state
+printAssumptionState ::
+  IsSymInterface sym =>
+  proxy sym ->
+  OverrideSim p sym ext r EmptyCtx UnitType (RegValue sym UnitType)
+printAssumptionState _proxy = do
+  ctx <- State.gets (Lens.view stateContext)
+  let hdl = printHandle ctx
+  CS.ovrWithBackend $ \bak -> liftIO $ do
+    let render =  PP.renderIO hdl . PP.layoutSmart PP.defaultLayoutOptions
+    let sym = CB.backendGetSym bak
+    state <- getBackendState bak
+    render (ppAssumptionState (Just sym) state)